lmkd: Fix shift-out-of-bounds UB
Decay logic introduced in aosp/1414468 ("lmkd: adjust thrashing dection
strategy") shifts the thrashing counter by the number of windows since
the last reset. That number can be large (it depends on the time between
consecutive pressure events), and shifting an int64_t by a value >= 64
is undefined behavior.

On both arm and x86, only the bottom 6 bits of the register holding the
shift amount are used, which means that the shift operation can produce
an incorrect result instead of decaying to 0.

Test: compile with ubsan enabled and check that lmkd doesn't crash
Change-Id: Idd317430dbc1cb998d7411c0b905084aa8b3c6ca
Signed-off-by: Tomislav Novak <tnovak@meta.com>