lmkd: Slightly reduce stats buffer size
Our buffer for LMK_KILL_OCCURRED_PACKET was oversized by 8 bytes,
because two of our 64-bit fields get encoded as 32-bit fields for
the packet.

This wasn't buggy, just a very small waste of space, and potentially
confusing for readers of the code.  We clean up the comments a bit
as well to make this clearer.

Bug: 405436516
Test: TreeHugger
Change-Id: I27ca953fc03d6c28c3cba62712dbb79c2300a007